% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement.R
\name{agreement}
\alias{agreement}
\title{Agreement between models}
\usage{
agreement(mod1, mod2, ref)
}
\arguments{
\item{mod1}{a (list) of models trained using the same algorithm}

\item{mod2}{a (list) of models trained using the same (different from
\code{mod1}) algorithm}

\item{ref}{(int) vector of correct classes}
}
\value{
a (list) with agreement metrics
}
\description{
Agreement between models
}
