% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom_dtm.R
\name{atom_dtm}
\alias{atom_dtm}
\title{Create a dtm from a corpus (TF weights)}
\usage{
atom_dtm(corpus, step = 500L)
}
\arguments{
\item{corpus}{(list) of documents, or a list of character vectors
each one reporting tokens from a document}

\item{step}{(num) integer value (default is 500L) used to broken the
procedure in parts of at maximum \code{step} documents each one.
This is to help to don't overflow the RAM.}
}
\value{
am object of class simple_triplet_matrix (from package
\code{slam}) representing a document-term matrix in which each row
is a document, each columns is a term (or token) and the content is
the simple frequencies of that terms in that document
}
\description{
\code{atom_dtm} take a corpus (list of documents), tokenized or not,
and create the Document-Term Matrix (DTM) stored as sparse, i.e.
using the algorithm of the simple triplet matrix (three indices i, j,
v, in which the indices i, j represent the row and the column
position of an entry and v represents its content). Moreover, for
compatibility reasons with some machine learning algorithm which use
another convention for sparsity, the indices are ordered and with
priority i, j too.
}
