% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Confidence interval for performance}
\usage{
ci(perf, alpha = 0.05)
}
\arguments{
\item{perf}{(list) of statistics for a model (e.g. CV)}

\item{alpha}{(num) level of significance for the confidence interval}
}
\value{
a data frame with the confidence interval for the model
}
\description{
Confidence interval for performance
}
