% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csr_dtm.R
\name{csr_dtm}
\alias{csr_dtm}
\title{csr from stm}
\usage{
csr_dtm(stm)
}
\arguments{
\item{stm}{simple triplet matrix (with row and columns name as
document and terms matrix)}
}
\value{
a matrix.csr
}
\description{
csr from stm
}
