% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csr_dtmTfIdf.r
\name{csr_dtmTfIdf}
\alias{csr_dtmTfIdf}
\title{csr with TFiDF from stm}
\usage{
csr_dtmTfIdf(stm)
}
\arguments{
\item{stm}{simple triplet matrix (with row and colums name as
document and terms matrix)}
}
\value{
a matrix.csr
}
\description{
csr with TFiDF from stm
}
