% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_perf.R
\name{cv_perf}
\alias{cv_perf}
\title{Cross-Validation performance}
\usage{
cv_perf(preds, statistics = list(acc = "acc"), dgt = 4)
}
\arguments{
\item{preds}{(list) of object of class prediction}

\item{statistics}{(list, default = list(acc = 'acc')) named list of
statistics of interest (using the convention of the package ROCR)}

\item{dgt}{(num) number of significant digits to use for the
performances}
}
\value{
a list of performance
}
\description{
\code{cv_perf} compute the required performance for Cross-validated
binary classification using by the ROCR package
}
