% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue}
\alias{glue}
\alias{glunique}
\title{Glue character strings}
\usage{
glue(texts, collapse = " ")

glunique(text)
}
\arguments{
\item{texts}{(chr) character vector to collapse}

\item{collapse}{(chr, default = ' ') character string to separate the
results}

\item{text}{(chr) character vector to collapse}
}
\value{
a (chr) vector
}
\description{
\code{glue} collapse a character vector in a single field eliminating
missing values.

\code{glunique} collapse a character vector in a single field,
eliminating missing values and duplicates elements (of the original
character vector)
}
