% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grams.R
\name{grams}
\alias{grams}
\alias{ngram}
\alias{bigram}
\alias{trigram}
\title{n-Gram creator}
\usage{
ngram(corpus, n_min = 1, n_max = 2, parallel = FALSE)

bigram(corpus, parallel = FALSE)

trigram(corpus, parallel = FALSE)
}
\arguments{
\item{corpus}{(list) of documents stored in a character vector of length 1}

\item{n_min}{(num) minimum number of words to include in the grams}

\item{n_max}{(num) maximum number of words to include into the grams}

\item{parallel}{(lgl) if \code{TRUE} perform the computation in parallel
using the \code{parallel} package functionality}
}
\value{
(list) of character vectors containing the nGrammed documents
}
\description{
a shortcuts for \code{ngram} using \code{n_min = n_max = 2}

a shortcuts for \code{ngram} using \code{n_min = n_max = 3}
}
