% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowering.R
\name{lowering}
\alias{lowering}
\title{Lowering}
\usage{
lowering(corpus)
}
\arguments{
\item{corpus}{(list) of (token of) documents}
}
\value{
a (list) of document all converted to lowercase
}
\description{
\code{lowering} mutate to lowercase every documents in a corpus provided as
a list of texts or a list of character vectors representing tokens
}
