% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltzostercode-package.R
\docType{package}
\name{mltzostercode-package}
\alias{mltzostercode}
\alias{mltzostercode-package}
\title{mltzostercode: Use of Machine Learning techniques for case-detection of
    Varicella Zoster using routinely collected textual ambulatory records}
\description{
Collect the function and the code used for the
    analyses reported in the paper "Use of Machine Learning techniques for
    case-detection of Varicella Zoster using routinely collected textual
    ambulatory records".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/mltzostercode}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/mltzostercode/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@unipd.it}

Authors:
\itemize{
  \item Paola Berchialla \email{paola.berchiall@unito.it}
  \item Ileana Baldi \email{ileana.baldi@unipd.it}
  \item Dario Gregori \email{dario.gregori@unipd.it} [funder]
}

Other contributors:
\itemize{
  \item Giulia Lorenzoni \email{giulia.lorenzoni@unipd.it} [contributor]
  \item Lara Tramontan \email{ltramontan@consorzioarsenal.it} [contributor]
  \item Antonio Scamarcia \email{a.scamarcia@sosepe.com} [data contributor]
  \item Luigi Cantarutti \email{l.cantarutti@sosepe.com} [data contributor]
  \item Carlo Giaquinto \email{carlo.giaquinto@unipd.it} [data contributor, funder]
}

}
\keyword{internal}
