% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonword.R
\name{nonword}
\alias{nonword}
\title{Eliminate non-words}
\usage{
nonword(corpus, numbers = FALSE)
}
\arguments{
\item{corpus}{(list) of documents, or a list of character vectors each one
reporting tokens from a document}

\item{numbers}{(lgl) if TRUE also numbers are removed (default FALSE)}
}
\value{
a (list) of document with non words removed
}
\description{
\code{nonword} eliminate everything is not an alphanumeric word/token
}
