% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc}
\alias{preproc}
\title{Preprocessing data}
\usage{
preproc(corpus)
}
\arguments{
\item{corpus}{(list) of documents, or a list of character vectors each one
reporting tokens from a document}
}
\value{
a list of document preprocessed
}
\description{
Preprocessing data
}
