% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem.R
\name{stem}
\alias{stem}
\title{Stem words}
\usage{
stem(corpus, language = "italian")
}
\arguments{
\item{corpus}{(list) of documents, or a list of character vectors each one
reporting tokens from a document}

\item{language}{(chr) language to pass to package \code{Snowballc} to perform
the stemming, default is 'italian'}
}
\value{
(list) of character vectors containing stemmed (tokens of the)
documents
}
\description{
Stem words
}
