% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_dns.R
\name{stm_dns}
\alias{stm_dns}
\title{Simple Triplet-Matrix to Dense matrix}
\usage{
stm_dns(dtm, memory = NULL)
}
\arguments{
\item{dtm}{simple triplet-matrix (document term matrix)}

\item{memory}{(num) the maximum amount of memory (RAM + virtual) the
system is required to reserve for the process. Default is to use
standard R option and so provide "memory.limit()"}
}
\value{
a full dense matrix
}
\description{
Simple Triplet-Matrix to Dense matrix
}
