% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{stopwords}
\alias{stopwords}
\alias{stopw}
\title{Remove stopwords}
\usage{
stopw(corpus, language = "italian", stopwords = NULL, stem = FALSE,
  parallel = FALSE)
}
\arguments{
\item{corpus}{(list) of documents, or a list of character vectors
each one reporting tokens from a document}

\item{language}{(chr) optional language to pass to package \code{tm}
if the parameter \code{stopwords} is left as default for
determination of the stopwords (default is 'italian')}

\item{stopwords}{(chr) optional vectors of stopwords, by default the
vector is provided by the \code{tm} package for the language
choosed by \code{language} (default is italian)}

\item{stem}{(lgl) if \code{TRUE} perform the stemming (using the
package \code{SnowballC}) of the stopwords before to remove them,
default is \code{FALSE}}

\item{parallel}{(lgl) if \code{TRUE} perform the computation in
parallel using the \code{parallel} package functionality}
}
\value{
(list) of documents pruned by the stopwords
}
\description{
Remove stopwords
}
