% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_varicella.R
\name{tag_varicella}
\alias{tag_varicella}
\title{Tag varicella}
\usage{
tag_varicella(corpus)
}
\arguments{
\item{corpus}{(list) of documents, or a list of character vectors
each one reporting tokens from a document}
}
\value{
a list of document with uniform tag for "varicella"
}
\description{
Tag varicella
}
