% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf4stm.R
\name{tfidf4stm}
\alias{tfidf4stm}
\title{Tf-IDf for stm DTM}
\usage{
tfidf4stm(stm, normalize = TRUE)
}
\arguments{
\item{stm}{an object of class simple_triplet_matrix representing a
DTM, in particular, with documents as row}

\item{normalize}{(lgl) if TRUE (default) the weights are normalized
wrt documents, i.e. row_sum equals 1 for every row.}
}
\value{
a \code{DocumentTermMatrix}, \code{simple_triplet_matrix}
}
\description{
Tf-IDf for stm DTM
}
