% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white.R
\name{white}
\alias{white}
\title{Strip White Spaces}
\usage{
white(corpus)
}
\arguments{
\item{corpus}{(list) of documents, or a list of character vectors each one
reporting tokens from a document}
}
\value{
a corpus (list)
}
\description{
\code{white} collapse all multiple "spaces" in a single space
}
